fix: add sign step
