+++ src/components/common/mint-step-modal/mint-step-modal.container.tsx
@@ -1,4 +1,5 @@
 import { StepStatus } from 'vwbl-sdk';
+import { VwblContainer } from '../../../container';
 import { MintStepModal as MintStepModalComponent } from './mint-step-modal';
 
 type Props = {
@@ -10,9 +11,12 @@ type Props = {
 };
 
 export const MintStepModal: React.FC<Props> = ({ isOpen, onClose, handleCancelClick, onMintClick, mintStep }: Props) => {
+  const { vwbl } = VwblContainer.useContainer();
+
   return (
     <MintStepModalComponent
       isOpen={isOpen}
+      signature={vwbl?.signature}
       onClose={onClose}
       mintStep={mintStep}
       handleMintStart={onMintClick}
