--- src/components/common/mint-step-modal/types.ts
@@ -1,14 +1,0 @@
-import { StepStatus } from 'vwbl-sdk';
-
-export type Props = {
-  isOpen: boolean;
-  mintStep: StepStatus[];
-  onClose: () => void;
-  handleCancelClick: () => void;
-  handleMintStart: () => void;
-};
-
-export type MintStepProps = {
-  mintStep: StepStatus[];
-  handleMintStart: () => void;
-};
